+++ demo/index.ts
@@ -1,8 +1,8 @@
-import { basicSetup, EditorState, EditorView } from '@codemirror/basic-setup';
+import { EditorState, EditorView } from '@codemirror/basic-setup';
 import { markdown, markdownLanguage } from '@codemirror/lang-markdown';
 import { theme } from './theme';
 import { indentWithTab } from '@codemirror/commands';
-import { keymap } from '@codemirror/view';
+import { highlightActiveLine, keymap } from '@codemirror/view';
 import { yaml } from '@codemirror/legacy-modes/mode/yaml';
 import { StreamLanguage } from '@codemirror/stream-parser';
 import {
@@ -15,11 +15,13 @@ import {
     lists,
     headingSlugField,
 } from '../src';
+import { defaultHighlightStyle } from '@codemirror/highlight';
 
 const editor = new EditorView({
     state: EditorState.create({
         extensions: [
-            basicSetup,
+            highlightActiveLine(),
+            defaultHighlightStyle,
             keymap.of([indentWithTab]),
             markdown({
                 base: markdownLanguage,