+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # @retronav/ixora
 
+## 0.1.3
+
+### Patch Changes
+
+-   5a90f6e: - Keep all plugin files seperate in production build, to improve modularity.
+    -   Introduce a default export consisting of all plugins.
+    -   Update dependencies and remove deprecated CodeMirror packages.
+    -   Remove the "YAML" prefix from all node names in the frontmatter plugin.
+
 ## 0.1.2
 
 ### Patch Changes